More code cleanups/refactoring. Moved audio.c (libavcodec based decoding stuff) into decoder.c, easier this way. Both Video and Audio threads are similar, and dispatcher for it is even more similar. Got rid of __int64 everywhere. Moved another qsort() helper into util.c include guards into sources.h and TSReader.h Replace a bunch of PID printings w/mask with FormatPIDMask() - this is still kinda ugly - need to revisit Refactored SI Tree Icon list - made them have proper names instead of magic numbers Got rid of link stubs.c Changed pcap homepage to current one (lol)

+++ parser.h
@@ -8,7 +8,7 @@ void ParseDVBSDTPacket(BYTE * pSectionPointer, int nPacketLength);
 BOOL ParseDCIIPMTTypeThing(BYTE * pSectionPointer, int nPacketLength, int nCurrentProgramNumber);
 BOOL ParseDCIIProgramNameMessage(BYTE * pSectionPointer, int nPacketLength, int nCurrentProgramNumber);
 BOOL ParsePMTPacket(BYTE * pSectionPointer, int nPacketLength, int nCurrentProgramNumber, int nPMTListenIndex);
-void ParseIPPacket(BYTE * pSectionPointer, int nPacketLength, int nPID, int nBufferNumber);
+void ParseIPPacket(BYTE * pSectionPointer, int nPacketLength, uint16_t nPID, int nBufferNumber);
 BOOL ParsePATPacket(BYTE * pSectionPointer, int nPacketLength);
 BOOL ParsePSIPPacket(BYTE * pSectionPointer, int nPacketLength);
 BOOL ParseCATPacket(BYTE * pSectionPointer, int nPacketLength);
